Rename MF packages
