+++ lenskit-integration-tests/src/it/eval/svd-identical/eval.groovy
@@ -35,14 +35,14 @@ import org.grouplens.lenskit.knn.item.model.*
 import org.grouplens.lenskit.knn.user.NeighborFinder
 import org.grouplens.lenskit.knn.user.SnapshotNeighborFinder
 import org.grouplens.lenskit.knn.user.UserUserItemScorer
-import org.grouplens.lenskit.mf.funksvd.FeatureCount
-import org.grouplens.lenskit.mf.funksvd.FunkSVDItemScorer
-import org.grouplens.lenskit.mf.funksvd.FunkSVDModelBuilder
-import org.grouplens.lenskit.mf.funksvd.FunkSVDUpdateRule
-import org.grouplens.lenskit.mf.svd.BiasedMFItemScorer
-import org.grouplens.lenskit.mf.svd.BiasedMFKernel
-import org.grouplens.lenskit.mf.svd.DotProductKernel
-import org.grouplens.lenskit.mf.svd.MFModel
+import org.lenskit.mf.funksvd.FeatureCount
+import org.lenskit.mf.funksvd.FunkSVDItemScorer
+import org.lenskit.mf.funksvd.FunkSVDModelBuilder
+import org.lenskit.mf.funksvd.FunkSVDUpdateRule
+import org.lenskit.mf.svd.BiasedMFItemScorer
+import org.lenskit.mf.svd.BiasedMFKernel
+import org.lenskit.mf.svd.DotProductKernel
+import org.lenskit.mf.svd.MFModel
 import org.grouplens.lenskit.transform.normalize.MeanCenteringVectorNormalizer
 import org.grouplens.lenskit.transform.normalize.UserVectorNormalizer
 import org.grouplens.lenskit.transform.normalize.VectorNormalizer
