--- lenskit-integration-tests/src/test/groovy/org/grouplens/lenskit/mf/funksvd/FunkSVDAccuracyTest.groovy
@@ -1,66 +1,0 @@
-/*
- * LensKit, an open source recommender systems toolkit.
- * Copyright 2010-2014 LensKit Contributors.  See CONTRIBUTORS.md.
- * Work on LensKit has been funded by the National Science Foundation under
- * grants IIS 05-34939, 08-08692, 08-12148, and 10-17697.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- */
-package org.grouplens.lenskit.mf.funksvd;
-
-import org.grouplens.lenskit.config.ConfigHelpers;
-import org.grouplens.lenskit.core.LenskitConfiguration;
-import org.grouplens.lenskit.iterative.IterationCount;
-import org.grouplens.lenskit.test.CrossfoldTestSuite;
-import org.grouplens.lenskit.util.table.Table
-import org.lenskit.api.ItemScorer
-import org.lenskit.baseline.BaselineScorer
-import org.lenskit.baseline.ItemMeanRatingItemScorer
-import org.lenskit.baseline.MeanDamping
-import org.lenskit.baseline.UserMeanBaseline
-import org.lenskit.baseline.UserMeanItemScorer;
-
-import static org.hamcrest.Matchers.closeTo;
-import static org.junit.Assert.assertThat;
-
-/**
- * Do major tests on the FunkSVD recommender.
- *
- * @author <a href="http://www.grouplens.org">GroupLens Research</a>
- */
-public class FunkSVDAccuracyTest extends CrossfoldTestSuite {
-    @SuppressWarnings("unchecked")
-    @Override
-    protected void configureAlgorithm(LenskitConfiguration config) {
-        ConfigHelpers.configure(config) {
-            bind ItemScorer to FunkSVDItemScorer
-            bind (BaselineScorer, ItemScorer) to UserMeanItemScorer
-            bind (UserMeanBaseline, ItemScorer) to ItemMeanRatingItemScorer
-            within (BaselineScorer, ItemScorer) {
-                set MeanDamping to 10
-            }
-            set FeatureCount to 25
-            set IterationCount to 125
-        }
-    }
-
-    @Override
-    protected void checkResults(Table table) {
-        assertThat(table.column("MAE.ByRating").average(),
-                   closeTo(0.74d, 0.025d))
-        assertThat(table.column("RMSE.ByUser").average(),
-                   closeTo(0.92d, 0.05d))
-    }
-}
