--- lenskit-svd/src/main/java/org/grouplens/lenskit/mf/funksvd/RuntimeUpdate.java
@@ -1,42 +1,0 @@
-/*
- * LensKit, an open source recommender systems toolkit.
- * Copyright 2010-2014 LensKit Contributors.  See CONTRIBUTORS.md.
- * Work on LensKit has been funded by the National Science Foundation under
- * grants IIS 05-34939, 08-08692, 08-12148, and 10-17697.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- */
-package org.grouplens.lenskit.mf.funksvd;
-
-import org.grouplens.grapht.annotation.DefaultNull;
-
-import javax.inject.Qualifier;
-import java.lang.annotation.*;
-
-/**
- * Qualifier for the update rule used at runtime.  This update rule is used by the {@linkplain
- * FunkSVDItemScorer item scorer} to train up user preferences at score time, to run on a more
- * updated version of their profile than was available at the last model build.  By default, no
- * score-time updating is done.
- *
- * @since 1.1
- */
-@Documented
-@Qualifier
-@DefaultNull
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PARAMETER, ElementType.METHOD})
-public @interface RuntimeUpdate {
-}
